Creates webpack output configuration module
